pyproject.toml
[project]
name = "gpttherapy"
version = "0.1.0"
description = "Email-driven turn-based AI therapy and storytelling system"
authors = [
    { name = "PromptExecution", email = "hello@promptexecution.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "pydantic>=2.5.0",
    "email-validator>=2.1.0",
    "python-dateutil>=2.8.0",
    "whenever>=0.7.0",
    "psutil>=7.0.0",
    "transitions>=0.9.3",
    "nanoid>=2.0.0",
    "structlog>=25.4.0",
    "python-decouple>=3.8",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff format
]


[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
disable_error_code = ["no-redef", "assignment", "arg-type", "var-annotated", "attr-defined", "index", "return-value", "annotation-unchecked"]

# Ignore specific missing imports for external dependencies
[[tool.mypy.overrides]]
module = [
    "pytest",
    "structlog",
    "whenever",
    "transitions",
    "nanoid",
    "decouple",
    "email_validator",
    "pydantic"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[dependency-groups]
dev = [
    "boto3-stubs>=1.39.15",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.5",
    "types-boto3>=1.39.15",
    "types-nanoid>=2.0.0.20240601",
]
